Moved AbstractTrainer to core (#1313)
* Tabular: Moved distill_utils to core, moved trainer utils to core

* Moved AbstractTrainer to core
