+++ core/src/autogluon/core/trainer/__init__.py
@@ -1,0 +1,1 @@
+from .abstract_trainer import AbstractTrainer
