+++ tabular/src/autogluon/tabular/trainer/__init__.py
@@ -1,1 +1,1 @@
-from .auto_trainer import *
+from .auto_trainer import AutoTrainer
